Restructured the dev tree folder
