+++ .gitignore
@@ -1,66 +1,17 @@
-#########################################
-#                                       #
-#       TheOutfiled's Ignore file       #
-
-##### File globs to match ####
-syntax: glob
-# Resharper
 _ReSharper.*
-# Visual Studio Stuff
-*.obj
-*.pdb
-*.user
-*.aps
-*.pch
-*.vspscc
-*.vssscc
-*_i.c
-*_p.c
-*.ncb
+[Bb]in/
+[Oo]bj/
 *.suo
-*.tlb
-*.tlh
-*.bak
+*.user
+*.userprefs
 *.cache
-*.ilk
-*.log
-*.lib
-*.sbr
-*.scc
-*.csproj.user
-obj/
-bin/
-[Dd]ebug*/
-[Rr]elease*/
-[Pp]ackages*/
-*.[Pp]ublish.xml
-# General
 *.orig
-*/Thumbs.db
-*/log.txt
-/Assets/
-/Build/
-*.DS_Store
-/packages/AutoMapper.3.0.0
-/packages/ClientDependency.1.7.1.2
-/packages/ClientDependency-Mvc.1.7.0.4
-/packages/HtmlAgilityPack.1.4.6
-/packages/ImageProcessor.1.8.7.0
-/packages/ImageProcessor.Web.3.2.1.0
-/packages/Lucene.Net.2.9.4.1
-/packages/Microsoft.AspNet.Mvc.4.0.30506.0
-/packages/Microsoft.AspNet.Mvc.FixedDisplayModes.1.0.1
-/packages/Microsoft.AspNet.Razor.2.0.30506.0
-/packages/Microsoft.AspNet.WebApi.4.0.30506.0
-/packages/Microsoft.AspNet.WebApi.Client.4.0.30506.0
-/packages/Microsoft.AspNet.WebApi.Core.4.0.30506.0
-/packages/Microsoft.AspNet.WebApi.WebHost.4.0.30506.0
-/packages/Microsoft.AspNet.WebPages.2.0.30506.0
-/packages/Microsoft.Net.Http.2.0.20710.0
-/packages/Microsoft.Web.Infrastructure.1.0.0.0
-/packages/MiniProfiler.2.1.0
-/packages/MySql.Data.6.6.5
-/packages/Newtonsoft.Json.6.0.2
-/packages/SharpZipLib.0.86.0
-/packages/UmbracoCms.Core.7.1.1
-/packages/xmlrpcnet.2.5.0
+Thumbs.db
+.DS_Store
+*.log
+.vs
+
+artifacts/
+src/packages/*/**
+build/[Cc]onfig/
+build/[Bb]in/
