--- Package.build.cmd
@@ -1,2 +1,0 @@
-C:\WINDOWS\Microsoft.NET\Framework\v4.0.30319\msbuild.exe Package.build.xml
-pause
